requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-colorfield==0.12.0
django-geoip2-extras==4.1
django-tailwind==3.8.0
geoip2==3.0.0
idna==3.10
maxminddb==2.6.3
pillow==11.1.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
